Fix scrolling on multi-monitor setups
WM_MOUSEWHEEL gives mouse coordinates in screen space, as signed 16-bit
numbers, but my code was treating the numbers as unsigned 16-bit
numbers. This caused the right split to scroll when trying to scroll the
left view when the am7 window was on a monitor to the left of the main monitor,
because that monitor would have negative screen coordinates.

+++ source/backend_win32.cpp
@@ -426,8 +426,8 @@ s64 my_window_proc(void *window_handle, s32 message, u64 w, s64 l)
             bool control = w & win32::MK_CONTROL;
             bool shift = w & win32::MK_SHIFT;
 
-            s32 sx = l & 0xffff;
-            s32 sy = (l >> 16) & 0xffff;
+            s32 sx = (s32) (s16) (l & 0xffff);
+            s32 sy = (s32) (s16) ((l >> 16) & 0xffff);
             win32::point Point = { sx, sy };
             win32::ScreenToClient(window_handle, &Point);
             s32 x = Point.X;
